requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Brotli==1.1.0
cupy-cuda12x==13.3.0
ffmpeg==1.4
gradio==4.44.0
numpy==1.24.4
openpyxl==3.1.5
Pillow==9.4.0
pydub==0.25.1
python-docx==1.1.2
torch
torchvision
tqdm==4.66.5
zstandard==0.23.0
pypdf2
python-docx 
pdfmerge
opencv-python

